Lezione_5/requirements.txt
openai
python-dotenv
tenacity
streamlit